pyproject.toml
[project]
name = "genai"
version = "0.1.0"
description = "Gen AI"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.8",
    "httpx>=0.28.1",
    "litellm>=1.53.3",
    "numpy>=2.2.2",
    "pandas>=2.2.3",
    "pgvector>=0.3.6",
    "psycopg2>=2.9.10",
    "pydantic>=2.10.6",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "scikit-learn>=1.6.1",
    "sentence-transformers>=3.4.1",
    "sqlalchemy>=2.0.37",
    "uvicorn>=0.34.0",
]
